Several unit tests added

+++ modules/test-tasks/src/apb/tests/utils/FileAssert.java
@@ -1,9 +1,6 @@
-package apb.tests.utils;
 
-import java.io.File;
 
-import junit.framework.Assert;
-import apb.utils.FileUtils;// Copyright 2008-2009 Emilio Lopez-Gabeiras
+// Copyright 2008-2009 Emilio Lopez-Gabeiras
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,33 +13,81 @@ import apb.utils.FileUtils;// Copyright 2008-2009 Emilio Lopez-Gabeiras
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License
+//
 
-// User: emilio
-// Date: Sep 5, 2009
-// Time: 8:53:25 PM
 
-public class FileAssert {
+package apb.tests.utils;
+
+import java.io.File;
+import java.util.Set;
+import java.util.TreeSet;
+
+import apb.utils.FileUtils;
+
+import junit.framework.Assert;
+
+import static java.util.Arrays.asList;
+
+public class FileAssert
+{
+    //~ Methods ..............................................................................................
+
     public static void assertDoesNotExist(File file)
     {
-        Assert.assertFalse("File: '" + file.getAbsolutePath() + " still exists.", file.exists());
+        Assert.assertFalse("File: '" + file.getAbsolutePath() + "' still exists.", file.exists());
     }
 
     public static void assertExists(File file)
     {
-        Assert.assertTrue("File: '" + file.getAbsolutePath() + " does not exist.", file.exists());
+        Assert.assertTrue("File: '" + file.getAbsolutePath() + "' does not exist.", file.exists());
     }
 
-    public static void assertDirEquals(File dir, File targetDir) {
-        File[] l = dir.listFiles();
-        for (File file1 : l) {
-            final File file2 = new File(targetDir, FileUtils.removePrefix(dir, file1));
-            assertFileEquals(file1, file2);
+    public static void assertDirEquals(File dir, File targetDir)
+    {
+        assertDirEquals(dir, targetDir, true);
+    }
+
+    public static void assertDirEquals(File dir, File targetDir, boolean compareContent)
+    {
+        File[] files = dir.listFiles();
+        File[] targetFiles = targetDir.listFiles();
+
+        if (files == null) {
+            assertExists(dir);
+        }
+        else if (targetFiles == null) {
+            assertExists(targetDir);
+        }
+        else {
+            Set<File> target = new TreeSet<File>(asList(targetFiles));
+
+            for (File file1 : files) {
+                final File file2 = new File(targetDir, FileUtils.removePrefix(dir, file1));
+
+                if (compareContent) {
+                    assertFileEquals(file1, file2);
+                }
+                else {
+                    Assert.assertEquals(file1.getName(), file2.getName());
+                }
+
+                target.remove(file2);
+            }
+
+            final boolean allFound = target.isEmpty();
+
+            if (!allFound) {
+                Assert.assertTrue("Extra files: " + target, allFound);
+            }
         }
     }
 
-    public static void assertFileEquals(File file1, File file2) {
+    public static void assertFileEquals(File file1, File file2)
+    {
         assertExists(file2);
-        final String msg = String.format("File content differs between '%s' and '%s'", file1, file2);
+        final String msg =
+            String.format("File content differs between '%s'.\n" + "                         and '%s'.",
+                          file1.getAbsolutePath(), file2.getAbsolutePath());
         Assert.assertTrue(msg, FileUtils.equalsContent(file1, file2));
     }
 }
